# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hoddmímir API",
    "version": "1.0.0",
    "description": "Authenticated, role-protected API for read projections and verified Hoddmímir configuration mutations."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "listAdministrationUsers",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Search" },
          { "$ref": "#/components/parameters/Enabled" }
        ],
        "responses": {
          "200": {
            "description": "A password-free user page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrationUserPage"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      },
      "post": {
        "operationId": "createAdministrationUser",
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CommandApplied" },
          "400": { "$ref": "#/components/responses/InvalidCommand" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "409": { "$ref": "#/components/responses/CommandConflict" },
          "422": { "$ref": "#/components/responses/SecurityCommandBlocked" },
          "503": { "$ref": "#/components/responses/SecurityCommandUnavailable" }
        }
      }
    },
    "/api/v1/admin/users/{id}": {
      "put": {
        "operationId": "updateAdministrationUser",
        "parameters": [
          { "$ref": "#/components/parameters/UserId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CommandApplied" },
          "400": { "$ref": "#/components/responses/InvalidCommand" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "409": { "$ref": "#/components/responses/CommandConflict" },
          "422": { "$ref": "#/components/responses/SecurityCommandBlocked" },
          "503": { "$ref": "#/components/responses/SecurityCommandUnavailable" }
        }
      }
    },
    "/api/v1/admin/users/{id}/disable": {
      "post": {
        "operationId": "disableAdministrationUser",
        "parameters": [
          { "$ref": "#/components/parameters/UserId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CommandApplied" },
          "400": { "$ref": "#/components/responses/InvalidCommand" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "409": { "$ref": "#/components/responses/CommandConflict" },
          "422": { "$ref": "#/components/responses/SecurityCommandBlocked" },
          "503": { "$ref": "#/components/responses/SecurityCommandUnavailable" }
        }
      }
    },
    "/api/v1/admin/users/{id}/roles": {
      "put": {
        "operationId": "replaceAdministrationUserRoles",
        "parameters": [
          { "$ref": "#/components/parameters/UserId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRolesCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CommandApplied" },
          "400": { "$ref": "#/components/responses/InvalidCommand" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "409": { "$ref": "#/components/responses/CommandConflict" },
          "422": { "$ref": "#/components/responses/SecurityCommandBlocked" },
          "503": { "$ref": "#/components/responses/SecurityCommandUnavailable" }
        }
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "listAdministrationRoles",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": {
          "200": {
            "description": "Closed roles and permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrationRolePage"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
    "/api/v1/admin/audit-events": {
      "get": {
        "operationId": "listAdministrationAuditEvents",
        "parameters": [
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/ActorUserId" },
          { "$ref": "#/components/parameters/AuditEventType" },
          { "$ref": "#/components/parameters/AuditOutcome" }
        ],
        "responses": {
          "200": {
            "description": "Append-only security audit events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrationAuditEventPage"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
    "/api/v1/admin/audit-events/{id}": {
      "get": {
        "operationId": "getAdministrationAuditEvent",
        "parameters": [{ "$ref": "#/components/parameters/AuditEventId" }],
        "responses": {
          "200": {
            "description": "One append-only audit event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrationAuditEvent"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "403": { "$ref": "#/components/responses/PermissionDenied" },
          "404": { "description": "Audit event not found." },
          "503": { "$ref": "#/components/responses/ReadModelUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hoddmimir_session"
      }
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/PageCursor" }
      },
...
      "Search": {
        "name": "search",
        "in": "query",
        "schema": { "type": "string", "minLength": 1, "maxLength": 190 }
      },
      "Enabled": {
        "name": "enabled",
        "in": "query",
        "schema": { "type": "boolean" }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
        }
      },
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "minLength": 43, "maxLength": 43 }
      },
      "UserId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
      "AuditEventId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
      "ActorUserId": {
        "name": "actorUserId",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/CanonicalUuid" }
      },
      "AuditEventType": {
        "name": "eventType",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/AuditEventType" }
      },
      "AuditOutcome": {
        "name": "outcome",
        "in": "query",
        "schema": { "type": "string", "enum": ["succeeded", "denied"] }
      }
    },
    "responses": {
...
      "PermissionDenied": {
        "description": "Permission or CSRF validation failed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthError" }
          }
        }
      },
      "InvalidQuery": {
        "description": "Unknown, malformed, incompatible, or out-of-bounds query.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
      "ReadModelUnavailable": {
        "description": "The read-only projection is temporarily unavailable without exposing database or runtime details.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [{ "$ref": "#/components/schemas/ReadModelUnavailableError" }, { "$ref": "#/components/schemas/MaintenanceError" }]
            }
          }
        }
      },
      "CommandApplied": {
        "description": "The command was applied or replayed idempotently.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ConfigurationCommandResult"
            }
          }
        }
      },
      "InvalidCommand": {
        "description": "The command envelope or closed payload is invalid.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/InvalidCommandError" }
          }
        }
      },
      "CommandConflict": {
        "description": "The expected revision or idempotency payload conflicts.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ConfigurationConflictError"
            }
          }
        }
      },
...
      "SecurityCommandBlocked": {
        "description": "The last-admin, self-lockout, or user-state invariant blocked the command.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SecurityCommandBlockedError"
            }
          }
        }
      },
      "SecurityCommandUnavailable": {
        "description": "The security administration store is temporarily unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [{ "$ref": "#/components/schemas/SecurityCommandUnavailableError" }, { "$ref": "#/components/schemas/MaintenanceError" }]
            }
          }
        }
      }
    },
    "schemas": {
...
      "MaintenanceError": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": {
          "type": "object", "additionalProperties": false, "required": ["code", "message"],
          "properties": {
            "code": { "const": "maintenance" },
            "message": { "const": "Hoddmímir wird gewartet." }
          }
        } }
      },
...
      "AuthError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "authentication_required",
                  "authentication_failed",
                  "permission_denied"
                ]
              }
            }
          }
        }
      },
      "RevisionCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedRevision"],
        "properties": {
          "expectedRevision": { "type": "integer", "minimum": 0 }
        }
      },
...
      "AdministrationRoleName": {
        "type": "string",
        "enum": ["admin", "viewer"]
      },
      "UserCreateCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expectedRevision",
          "username",
          "displayName",
          "password",
          "roles"
        ],
        "properties": {
          "expectedRevision": { "type": "integer", "const": 0 },
          "username": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9._-]{2,63}$"
          },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 190 },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          },
          "roles": {
            "type": "array",
            "minItems": 1,
            "maxItems": 2,
            "uniqueItems": true,
            "items": { "$ref": "#/components/schemas/AdministrationRoleName" }
          }
        }
      },
      "UserUpdateCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedRevision", "displayName"],
        "properties": {
          "expectedRevision": { "type": "integer", "minimum": 1 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 190 },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          }
        }
      },
      "UserRolesCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedRevision", "roles"],
        "properties": {
          "expectedRevision": { "type": "integer", "minimum": 1 },
          "roles": {
            "type": "array",
            "minItems": 1,
            "maxItems": 2,
            "uniqueItems": true,
            "items": { "$ref": "#/components/schemas/AdministrationRoleName" }
          }
        }
      },
...
      "ConfigurationCommandResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "revision"],
        "properties": {
          "status": { "type": "string", "enum": ["applied", "replayed"] },
          "revision": { "type": "integer", "minimum": 1 }
        }
      },
      "InvalidCommandError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": { "code": { "const": "invalid_request" } }
          }
        }
      },
      "ConfigurationConflictError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "currentRevision"],
            "properties": {
              "code": { "const": "revision_conflict" },
              "currentRevision": { "type": "integer", "minimum": 0 }
            }
          }
        }
      },
...
      "CanonicalUuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      },
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{6}Z$"
      },
...
      "PageCursor": {
        "type": "string",
        "minLength": 1,
        "maxLength": 2048,
        "pattern": "^[A-Za-z0-9_-]+$"
      },
...
      "PageMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": ["limit", "count", "hasMore", "nextCursor"],
        "properties": {
          "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
          "count": { "type": "integer", "minimum": 0, "maximum": 100 },
          "hasMore": { "type": "boolean" },
          "nextCursor": {
            "oneOf": [
              { "$ref": "#/components/schemas/PageCursor" },
              { "type": "null" }
            ]
          }
        }
      },
...
      "AdministrationUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "username",
          "displayName",
          "enabled",
          "revision",
          "roles",
          "createdAt",
          "updatedAt",
          "disabledAt",
          "lastLoginAt"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "username": { "type": "string", "minLength": 3, "maxLength": 64 },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 190 },
          "enabled": { "type": "boolean" },
          "revision": { "type": "integer", "minimum": 1 },
          "roles": {
            "type": "array",
            "maxItems": 2,
            "uniqueItems": true,
            "items": { "$ref": "#/components/schemas/AdministrationRoleName" }
          },
          "createdAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "updatedAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "disabledAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "lastLoginAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          }
        }
      },
      "AdministrationUserPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/AdministrationUser" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
      "AdministrationRole": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "displayName", "permissions"],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "name": { "$ref": "#/components/schemas/AdministrationRoleName" },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 64 },
          "permissions": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "inventory.read",
                "backup_configuration.manage",
                "backup_operations.manage",
                "audit.read",
                "security.manage"
              ]
            }
          }
        }
      },
      "AdministrationRolePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/AdministrationRole" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
      "AuditEventType": {
        "type": "string",
        "enum": [
          "first_admin_created",
          "user_created",
          "user_updated",
          "user_disabled",
          "role_assigned",
          "role_removed",
          "login_succeeded",
          "login_failed",
          "session_created",
          "session_revoked",
          "target_created",
          "target_updated",
          "target_enabled",
          "target_disabled",
          "policy_created",
          "policy_updated",
          "policy_enabled",
          "policy_disabled",
          "selection_upserted",
          "selection_disabled",
          "guest_override_upserted",
          "guest_override_disabled",
          "connection_created",
          "connection_updated",
          "connection_enabled",
          "connection_disabled",
          "endpoint_created",
          "endpoint_updated",
          "endpoint_disabled",
          "credential_rotated"
          ,"manual_backup_requested",
          "backup_cancel_requested"
        ]
      },
      "AdministrationAuditEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "occurredAt",
          "actorUserId",
          "actorSessionId",
          "eventType",
          "outcome",
          "subjectType",
          "subjectId",
          "reasonCode",
          "correlationId"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/CanonicalUuid" },
          "occurredAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "actorUserId": {
            "oneOf": [
              { "$ref": "#/components/schemas/CanonicalUuid" },
              { "type": "null" }
            ]
          },
          "actorSessionId": {
            "oneOf": [
              { "$ref": "#/components/schemas/CanonicalUuid" },
              { "type": "null" }
            ]
          },
          "eventType": { "$ref": "#/components/schemas/AuditEventType" },
          "outcome": { "type": "string", "enum": ["succeeded", "denied"] },
          "subjectType": {
            "oneOf": [
              {
                "type": "string",
                "enum": [
                  "user",
                  "session",
                  "role",
                  "target",
                  "policy",
                  "selection",
                  "guest_override",
                  "connection"
                ]
              },
              { "type": "null" }
            ]
          },
          "subjectId": {
            "oneOf": [
              { "$ref": "#/components/schemas/CanonicalUuid" },
              { "type": "null" }
            ]
          },
          "reasonCode": {
            "oneOf": [
              { "type": "string", "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$" },
              { "type": "null" }
            ]
          },
          "correlationId": { "$ref": "#/components/schemas/CanonicalUuid" }
        }
      },
      "AdministrationAuditEventPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page"],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/AdministrationAuditEvent" }
          },
          "page": { "$ref": "#/components/schemas/PageMetadata" }
        }
      },
      "SecurityCommandBlockedError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "blockers"],
            "properties": {
              "code": { "const": "security_invariant_blocked" },
              "blockers": {
                "type": "array",
                "minItems": 1,
                "maxItems": 1,
                "items": {
                  "type": "string",
                  "enum": [
                    "user_missing",
                    "user_disabled",
                    "self_lockout",
                    "last_active_admin"
                  ]
                }
              }
            }
          }
        }
      },
      "SecurityCommandUnavailableError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code"],
            "properties": {
              "code": { "const": "security_administration_unavailable" }
            }
          }
        }
      },
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": { "const": "invalid_query" },
              "message": { "const": "The query is invalid." }
            }
          }
        }
      },
      "ReadModelUnavailableError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": { "const": "read_model_unavailable" },
              "message": {
                "const": "The read model is temporarily unavailable."
              }
            }
          }
        }
      }
    }
  }
}